requirements.txt
dask==2021.10.0
easydict==1.9
google-auth==2.6.0
google-auth-oauthlib==0.4.6
hydra-core==1.1.1
imageio==2.16.0
matplotlib==3.5.1
numpy==1.21.2
pandas==1.4.1
Pillow==9.0.1
plotly==5.6.0
pytorch-lightning==1.5.10
requests==2.27.1
requests-oauthlib==1.3.1
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.10.2
torchaudio==0.10.2
torchmetrics==0.7.2
torchvision==0.11.3
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.8
wandb==0.12.10
